+++ codegen/src/test/java/org/web3j/codegen/SolidityFunctionWrapperGeneratorTest.java
@@ -49,7 +49,7 @@ public class SolidityFunctionWrapperGeneratorTest extends TempFileProvider {
         super.setUp();
 
         URL url = SolidityFunctionWrapperGeneratorTest.class.getResource("/solidity");
-        solidityBaseDir = url.getPath();
+        solidityBaseDir = new File(url.toURI()).getAbsolutePath();
     }
 
     @Test
@@ -298,6 +298,8 @@ public class SolidityFunctionWrapperGeneratorTest extends TempFileProvider {
                                         "java",
                                         inputFileName + ".java"),
                                 File.separator));
+        System.out.println("EXPECTED: " + fileExpected.getAbsolutePath());
+        System.out.println("ACTUAL:   " + fileActual.getAbsolutePath());
         assertEquals(
                 new String(Files.readAllBytes(fileExpected.toPath())).replaceAll("(\r\n|\n)", ""),
                 new String(Files.readAllBytes(fileActual.toPath())).replaceAll("(\r\n|\n)", ""));